requirements.txt

streamlit
Pillow
ultralytics
